handling base case

+++ helpers/helpers.go
@@ -6,18 +6,18 @@ import (
 	"math"
 )
 
-func GetHash(key string, m int) int {
+func GetHash(key string, m int) int64 {
 	h := sha1.New()
 	h.Write([]byte(key))
 	hash := h.Sum(nil)
 	hashStr := fmt.Sprintf("%x", hash)
 
-	sum := 0
+	sum := int64(0)
 
 	for _, char := range hashStr {
-		asciiValue := int(char)
+		asciiValue := int64(char)
 		sum += asciiValue
 	}
 
-	return sum%int(math.Pow(2.0, float64(m)))
+	return sum%int64(math.Pow(2.0, float64(m)))
 }
